Some minor cleanup of the code (#13)
* lint: Arn -> ARN in types

* Remove hardcoded profile name from aws session creation.

There's no need to hardcode this in the code. The following command works just
as easily without the need for any such config:

AWS_PROFILE=event-handler AWS_REGION=us-east-1 ./bin/local/ecs-event-stream-handler

* Move logger to its own package.

Also get rid of the xml file and load from a const instead.

Another modification is to disable the fille logger for now
as the path var/output/... doesn't exist on a host and won't
work anyway. Use console logger for now. We should revisit
this in the future.

* Clean up logs and error messages in consumer and the stores.

Modified some error messages and logs to be more user-friendly.
Using the String() function for instance and task events to print
more information about what was added to the store.
Also using aws provided dereferencing methods instead of doing
it directly. The StringValue(), IntValue() and BoolValue()
methods are much safer than pointer dereferencing op.