Fix scheduler behavior to handle completed deployments correctly
Currently when deployments are created they end up in 'pending' state.
Scheduler picks up the 'pending' deployment and starts tasks. If all the
tasks are in running state then deployment is marked as 'completed'.
Subsequent runs of scheduler cannot continue because there is no
in-progress deployment. This fix will make sure that scheduler will pick
up most recent completed deployment when there is no in-progress
deployment.