Add the skeleton of an AWS API Gateway + Lambda frontend service
This adds a Spring MVC/REST web service, running as a Lambda function,
fronted by API Gateway. See `docs/frontend.md` for an overview of how
this works

The main moving parts are:
- Spring controller from which Swagger model is derived: `com.amazonaws.blox.frontend.controllers.EnvironmentController`
- Serverless Application Model template: `frontend-infrastructure/api/template.yml`
- Generated Swagger model: `frontend-service/api/swagger.yml`

There is some custom build logic in `buildSrc` that takes care of
generating a Swagger model from the Spring controller, and to add all
the necessary extensions to make it work with API Gateway.

This commit does not include the client that gets generated from the
API, for ease of reviewing.