Fix log4j configuration for end-to-end tests
Log4j wasn't actually working in the integ tests, claiming that it
couldn't find log4j.properties. I tried to fix it by copying the default
config from the docs, which fixed it. Turns out, there was a missing
comma.

The default config has a slightly nicer format, and some helpful
comments, so this keeps it.