Implement efficient cluster state snapshots in ECSStateClient (#275)
* Use lists instead of maps for tasks/instances in state snapshot

The original idea was to index tasks/instances by ID once in the
Manager, instead of having to reindex in each scheduler. However, in
practice we want to index tasks by instance id (not task id), so it's
more flexible to use lists instead.

* Make sure SingleTaskScheduler actually only starts one task

This requires persisting something in the taskGroup, so that we know
whether the scheduler started the task, or if it's unrelated.

* Implement efficient cluster state snapshots in ECSStateClient

This uses an `ECSAsyncClient` to implement the `ECSState` interface by
concurrently calling describe for sequential pages of list responses.

The integration test is very basic, but we may be able to build on it
for further integration tests that require a test cluster.

See #261 for details.