Wire up Data Service to Scheduling Manager using JSON-RPC (#281)
In order to use a single Lambda function for the DataService, we need
some form of RPC over the Lambda input/output streams. The jsonrpc4j
project supports this pretty well.

- `json-rpc-lambda-client` wraps a LambdaAsyncClient to call a lambda
  function with JSON-RPC formatted request and response payloads. It
  uses jsonrpc4j's Proxy support to dynamically generate a client for
  any given Java interface.
- `json-rpc-lambda-server` implements `RequestStreamHandler` and plumbs
  it into the Input-/OutputStream based JsonRpcServer.

This deploys the DataService as a lambda function that accepts JSON-RPC
payloads, and adds a JSON-RPC client to the Scheduling Manager in order
to call the DataService.

Right now, this wires up the DataService's Lambda function into the
scheduler using Cloudformation's ImportValue function. While this is nice
for development, we'll probably want to figure out a better long term
way to do this.

Implements #227