Clean up Gradle build scripts, move config to root project (#287)
* Upgrade Gradle to 4.3

The new version of gradle fixes some issues with the new `plugins`
directive. In particular, it's now possible to use it to activate
plugins in subprojects that have already been declared in the root
project.

* Clean up Gradle build scripts, move config to root project

This moves as much of our repeated subproject configuration as makes
sense to the root project, and versions more of the most common
dependencies using the dependencyManagement section.

This also consolidates all plugin versions in the root project using the
plugins {} block, and only applies them in subprojects that need them.