Use a common prefix to namespace resources created in a single scenario (#335)
This changes the InputCreator to prefix the same UUID to all resources
it creates, and sets up the DataServiceSteps as a DirtiesContext. The
effect of this is that all resources that are created in the same
scenario run will have the same prefix, and that prefix will be
different from all other scenario runs.

This lets us run the tests more than once, without needing to clean up
data.