Use a GSI query instead of scan to list clusters by accountId/name (#353)
This uses a GSI to power listClusters queries that have to filter by
accountId and clusterNamePrefix.

Unfortunately, it seems like using a separate POJO for the GSI
(EnvironmentClusterIndexDDBRecord) doesn't work as the docs say it does.
In particular, trying to do any queries using that POJO fails due to a
missing hash key attribute annotation. Therefore this change moves the
indexing annotations back into EnvironmentDDBRecord.

Also unfortunately, DDB does not deduplicate records in a GSI by
hash/range key, so we still have one record per Environment in the
cluster index, and we have to manually deduplicate the values we return.

To make testing easier, the unit tests for listClusters use DDBLocal so
that we don't have to construct accurate mocks for the DDB API's
behaviour.  For more testing convenience, this also introduces Test Data
Builders (see [this article][1] for definition) to make it easier to
construct valid model POJOs for our tests.

[1]: http://blog.codeleak.pl/2014/06/test-data-builders-and-object-mother.html