Fix NullPointerException for inactive environments in scheduler handler (#373)
The scheduler handler incorrectly assumed that an environment will
always have an active environment revision. This is not true for
newly-created environments, and would cause scheduler invocations to
fail with a NullPointerException.

Instead, it now returns an empty SchedulerOutput as soon as it's clear
that the environment has no active revisions.