Fix the E2E test for codebuild to let deploy first (#378)
The gradle task order does not imply execution order, therefore the
default endpoint for testEndToEnd is null as deployment output file is
not ready yet.

For local development, it won't be a problem as the deployment output
file should exist already. We make the code build as two steps so we
could have default endpoint ready before run test task.